Update word list and account for multiple sol'ns
Updated the wordlist to upper case.

Also changed the unscrambled word from a string to a List<string> because it is possible that a single word may have an anagram such as "CHEATER" and "TEACHER".  Now, both solutions work and the initial tile randomization is checked so neither word will appear in order initially.